+++ src/core/component/directives/aria/roles-engines/index.ts
@@ -1,1 +1,18 @@
-export * from 'core/component/directives/aria/roles-engines/dialog';
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+export { default as dialog } from 'core/component/directives/aria/roles-engines/dialog';
+export { default as tablist } from 'core/component/directives/aria/roles-engines/tablist';
+export { default as tab } from 'core/component/directives/aria/roles-engines/tab';
+export { default as tabpanel } from 'core/component/directives/aria/roles-engines/tabpanel';
+export { default as controls } from 'core/component/directives/aria/roles-engines/controls';
+export { default as combobox } from 'core/component/directives/aria/roles-engines/combobox';
+export { default as listbox } from 'core/component/directives/aria/roles-engines/listbox';
+export { default as option } from 'core/component/directives/aria/roles-engines/option';
+export { default as tree } from 'core/component/directives/aria/roles-engines/tree';
+export { default as treeitem } from 'core/component/directives/aria/roles-engines/treeitem';