+++ src/core/component/directives/aria/roles-engines/tabpanel.ts
@@ -1,0 +1,22 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+import AriaRoleEngine from 'core/component/directives/aria/interface';
+
+export default class TabpanelEngine extends AriaRoleEngine {
+	init(): void {
+		const
+			{el, binding} = this.options;
+
+		el.setAttribute('role', 'tabpanel');
+
+		if (binding.value?.labelledby == null) {
+			Object.throw('Tabpanel aria directive expects "label" or "labelledby" value to be passed');
+		}
+	}
+}