+++ src/traits/i-access/const.ts
@@ -1,0 +1,2 @@
+export const
+	FOCUSABLE_SELECTOR = '[tabindex]:not([disabled]), a:not([disabled]), button:not([disabled]), input:not([disabled]), select:not([disabled]), textarea:not([disabled])';