+++ src/base/b-tree/CHANGELOG.md
@@ -9,6 +9,15 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2022-??-??)
+
+#### :rocket: New Feature
+
+* Added `v-aria` directive
+* Added a new prop `vertical`
+* Added `changeFoldedMod`
+* Now the component derive `iAccess`
+
 ## v3.0.0-rc.164 (2021-03-22)
 
 #### :house: Internal